Clarify failing unit tests that #8 uncovered
Beyond the problem with prefix '-', the Hedgehog generator found another
problem with '.1' being parsed prematurely as the 'Identity' jq filter.

Problem 1:

The parser needs to address that '.<anything>' could be 'Identity', or
any of the other combinators listed in Parser's 'dotExpr' sub-parser
*including* the currently overshadowed '.1' of the 'number' sub-parser.

Until this is resolved, I'll leave the outcome of parsing '-.1' up in
the air.

In summary, this commit highlights the two problems:
 - '.1' cannot be determined as 'NumLit 0.1' with lookahead 1.
 - '-.1' should be either 'Neg (NumLit .1)' or 'Numlit (-0.1)'