[#14] Handle '.1', uncover some generator bugs
A type 'Jq.Number.JqNumber' is introduced. For now it is reduced to a
'Scientific' in the AST, but it can be used for structure-preserving
operations on numbers later on.

LL(1) ambiguity between '.foo' and '.1' is resolved for now by trying,
with backtracking, to parse a number before trying to parse any dot
expression.

The 'scientific' parser is replaced with a custom 'jqNumber' parser.

A bug in 'scientificNumberGen' was addressed that caused it to generate
"1e+-1". Its range is changed to an exponent of 309, since this is a
point where hs-jq's arbitrary-precision numbers will deviate from jq's
doubles. (See: 'jq -n 1e309', or other examples in #4.)

The 'hprop_NumLit' property is rephrased to use 'evalEither'. Since the
parser, for now, parses negative numbers as 'Neg (NumLit n)', we expect
this in the property.

Because the 'Read Scientific' instance does not allow '.1', the property
patches this by prepending a '0'.

Currently unhandled: "0." fails in parseHaskell (expected). This is not
valid JSON number syntax, and aeson rejects this syntax as well. We deal
with it later.