[#1] Test parser on jq's test suite
This imports the main test suite mentioned in #1. There are three other
test suites, base64.test, onig.test and optional.test that we'll deal
with later.

Currently, only the parser is tested, and we only verify whether it
thinks the parse succeeded or failed, and not if it actually parsed
something correctly. This can still be used as a progress bar on the
coverage (but not the correctness).